Add files via upload
Regression Analysis is a statistical process for estimating the relationships between the dependent variables or criterion variables and one or more independent variables or predictors. Regression analysis explains the changes in criterions in relation to changes in select predictors.  Three major uses for regression analysis are determining the strength of predictors, forecasting an effect, and trend forecasting. The Repository Contains Source code in Python as well as R-Programming. This Repository contains Source code of 6 Different Regression Techniques. 
1) Simple Linear Regression 
2) Multiple Linear Regression
3) Polynomial Regression
4) Support Vector Regression
5) Decision Tree Regression
6) Random forest Regression 
Refer to the Dataset Repository to get the dataset for all the models and algorithms.